+++ src/components/footer/Footer.astro
@@ -1,6 +1,6 @@
 ---
 import Logo from '../../assets/logo.svg';
-import {footerSections} from '../../constants';
+import {footerSections} from '../../config';
 
 const footerClasses = {
   'w-full bg-zinc-200 py-8 px-4': true,
