+++ src/components/nav/Navbar.astro
@@ -2,7 +2,7 @@
 import '@/styles/global.css';
 import Logo from '@/assets/logo.svg';
 import NavItems from './NavItems.astro';
-import {navItems} from '../../constants';
+import {navItems} from '../../config';
 import NavBarSearch from '../NavBarSearch.astro';
 ---
