+++ src/components/side-bar/SideBar.astro
@@ -1,6 +1,6 @@
 ---
-import Menu from './Menu.astro';
-import {sidebarItems} from '../constants/sidebar';
+import Menu from '../menu/Menu.astro';
+import {sidebarItems} from '../../config';
 
 const sidebarClasses = {
   'w-full': true,
