+++ src/components/toc/Toc.astro
@@ -1,7 +1,7 @@
 ---
 import type {MarkdownHeading} from 'astro';
 import TreeTocNode from './TreeTocNode.astro';
-import {buildTocTree} from '../lib/buildTocTree';
+import {buildTocTree} from '../../lib/buildTocTree';
 
 interface Props {
   toc: MarkdownHeading[];
