+++ src/components/toc/TreeTocNode.astro
@@ -1,5 +1,5 @@
 ---
-import type {TreeTocNode} from '../interface';
+import type {TreeTocNode} from '../../interface';
 
 interface Props {
   node: TreeTocNode;
