+++ src/config/index.ts
@@ -1,2 +1,3 @@
 export * from './nav';
 export * from './footer';
+export * from './sidebar';
