+++ src/config/sidebar.ts
