+++ src/interface/index.ts
@@ -7,6 +7,7 @@ export interface TreeTocNode extends MarkdownHeading {
 export interface NavItemProps {
   text: string;
   link: string;
+  isActive?: boolean;
 }
 
 export type NavItemsProps = NavItemProps[];
