+++ src/pages/doc/Document.astro
@@ -1,8 +1,8 @@
 ---
 import '@/styles/global.css';
 import type {MarkdownHeading} from 'astro';
-import Toc from './Toc.astro';
-import SideBar from './SideBar.astro';
+import Toc from '@/components/toc/Toc.astro';
+import SideBar from '@/components/SideBar.astro';
 
 interface Props {
   headings: MarkdownHeading[];
